Refactor feature keys
